requirements.txt
streamlit
python-dotenv
langchain
pypdf
faiss-cpu
InstructorEmbedding
sentence-transformers
huggingface_hub